spaced files out for readiblity, renamed files, added some error handling
