+++ dummy_data/Sean_test/cos_sim.py
